Reorganize the Complaint DTOs
