+++ src/AppServices/Complaints/CommandDto/ComplaintAssignDto.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Cts.AppServices.Complaints;
+namespace Cts.AppServices.Complaints.CommandDto;
 
 public record ComplaintAssignDto(int ComplaintId)
 {
