+++ src/AppServices/Complaints/CommandDto/ComplaintClosureDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Cts.AppServices.Complaints;
+﻿namespace Cts.AppServices.Complaints.CommandDto;
 
 // Used for approving/closing, reopening, deleting, and restoring complaints.
 public record ComplaintClosureDto(int ComplaintId)
