+++ src/AppServices/Complaints/Permissions/ComplaintUpdatePermissionsHandler.cs
@@ -1,4 +1,4 @@
-﻿using Cts.AppServices.Complaints.Dto.Command;
+﻿using Cts.AppServices.Complaints.CommandDto;
 using Cts.AppServices.Permissions.Helpers;
 using Cts.Domain;
 using Cts.Domain.Entities.Complaints;
