+++ src/AppServices/Complaints/QueryDto/ComplaintSearchEnums.cs
@@ -2,7 +2,7 @@ using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 
-namespace Cts.AppServices.Complaints.Dto.Query;
+namespace Cts.AppServices.Complaints.QueryDto;
 
 [JsonConverter(typeof(JsonStringEnumConverter))]
 public enum SortBy
