+++ src/AppServices/Complaints/QueryDto/ComplaintSearchResultDto.cs
@@ -1,7 +1,7 @@
 ﻿using Cts.AppServices.Staff.Dto;
 using Cts.Domain.Entities.Complaints;
 
-namespace Cts.AppServices.Complaints.Dto.Query;
+namespace Cts.AppServices.Complaints.QueryDto;
 
 public record ComplaintSearchResultDto
 {
