+++ src/WebApp/Pages/Public/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "{handler?}"
-@using Cts.AppServices.Complaints.Dto.Query
+@using Cts.AppServices.Complaints.QueryDto
 @using Cts.WebApp.Pages.Shared.DisplayTemplates
 @using Cts.WebApp.Platform.PageModelHelpers
 @model IndexModel
