+++ src/WebApp/Pages/Shared/_ComplaintSummaryPartial.cshtml
@@ -1,6 +1,6 @@
 ﻿@using Cts.WebApp.Pages.Shared.DisplayTemplates
 @using GaEpd.AppLibrary.Extensions
-@model Cts.AppServices.Complaints.Dto.Query.ComplaintViewDto
+@model Cts.AppServices.Complaints.QueryDto.ComplaintViewDto
 
 <div class="container small">
     <dl class="row">
