+++ src/WebApp/Pages/Staff/Complaints/Add.cshtml
@@ -1,6 +1,6 @@
 @page
 @using Cts.AppServices.Attachments
-@using Cts.AppServices.Complaints.Dto.Command
+@using Cts.AppServices.Complaints.CommandDto
 @using Cts.Domain.ValueObjects
 @using Cts.WebApp.Pages.Shared.EditorTemplates
 @using Cts.WebApp.Platform.PageModelHelpers
