+++ src/WebApp/Pages/Staff/Complaints/Assign.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "{id:int?}"
-@using Cts.AppServices.Complaints
+@using Cts.AppServices.Complaints.CommandDto
 @using Cts.WebApp.Pages.Shared.EditorTemplates
 @using Cts.WebApp.Platform.PageModelHelpers
 @using Microsoft.AspNetCore.Mvc.TagHelpers
