+++ src/WebApp/Pages/Staff/Complaints/DeleteAction.cshtml.cs
@@ -1,8 +1,8 @@
 ﻿using Cts.AppServices.ComplaintActions;
 using Cts.AppServices.ComplaintActions.Dto;
 using Cts.AppServices.Complaints;
-using Cts.AppServices.Complaints.Dto.Query;
 using Cts.AppServices.Complaints.Permissions;
+using Cts.AppServices.Complaints.QueryDto;
 using Cts.WebApp.Models;
 using Cts.WebApp.Platform.PageModelHelpers;
 using Microsoft.AspNetCore.Authorization;
