+++ src/WebApp/Pages/Staff/Complaints/Details.cshtml.cs
@@ -4,8 +4,8 @@ using Cts.AppServices.Attachments.Dto;
 using Cts.AppServices.ComplaintActions;
 using Cts.AppServices.ComplaintActions.Dto;
 using Cts.AppServices.Complaints;
-using Cts.AppServices.Complaints.Dto.Query;
 using Cts.AppServices.Complaints.Permissions;
+using Cts.AppServices.Complaints.QueryDto;
 using Cts.AppServices.Permissions;
 using Cts.AppServices.Staff;
 using Cts.WebApp.Models;
