+++ tests/AppServicesTests/Complaints/Create.cs
@@ -1,6 +1,6 @@
 ﻿using Cts.AppServices.Attachments;
 using Cts.AppServices.Complaints;
-using Cts.AppServices.Complaints.Dto.Command;
+using Cts.AppServices.Complaints.CommandDto;
 using Cts.AppServices.UserServices;
 using Cts.Domain.Entities.Complaints;
 using Cts.Domain.Entities.Concerns;
