+++ tests/AppServicesTests/Complaints/Search.cs
@@ -1,6 +1,6 @@
 ﻿using Cts.AppServices.Attachments;
 using Cts.AppServices.Complaints;
-using Cts.AppServices.Complaints.Dto.Query;
+using Cts.AppServices.Complaints.QueryDto;
 using Cts.AppServices.UserServices;
 using Cts.Domain.Entities.Complaints;
 using Cts.Domain.Entities.Concerns;
